+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `partial-recall-steal` to take another memory's buffer.
+
 ### Changed
 
 - The defaults for custom variables have been vastly increased.